(refactor): create util file for routes and validators
